Peel out NavItem to separate file
