+++ src/helpers/dom.js
@@ -51,3 +51,15 @@ export const findActiveNode = (possibleNodes, isScrollingDown) => {
     return bottom > topEdge && bottom < window.innerHeight;
   });
 };
+
+export const isInViewport = (elem) => {
+  const bounding = elem.getBoundingClientRect();
+  return (
+    bounding.top >= 0 &&
+    bounding.left >= 0 &&
+    bounding.bottom <=
+      (window.innerHeight || document.documentElement.clientHeight) &&
+    bounding.right <=
+      (window.innerWidth || document.documentElement.clientWidth)
+  );
+};
